+++ example/index.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const { Issuer } = require('..');
 
 const {