+++ test/tokenset/tokenset.test.js
@@ -1,8 +1,6 @@
-'use strict';
-
 const sinon = require('sinon');
 const base64url = require('base64url');
-const expect = require('chai').expect;
+const { expect } = require('chai');
 const TokenSet = require('../../lib/token_set');
 const now = require('../../lib/unix_timestamp');